Change from shift register to FIFO for storing raw image rows
The methodology used before forced the rest of the project to capture
images of fixed row size. The usage of a FIFO allows to dynamically
change the row size during execution without affecting the rest  of the
project. The FIFO is implemented using a generic Altera megafunction.

The output of the raw2rgb component is now simplified. Before, it
changed the RGB output on each cycle, despite the 'data valid' bit is
set to 1 only on a fourth of all cycles, namely when the ongoing pixel
is on an odd row and an odd column.

The assignment of each pixel components values was also clarified. As
the camera configuration includes a mirror operation on both image
dimensions, the pixel array structure is an inverted Bayer pattern.