Add a ZMQ socket for receiving speed set points
A poller is initialized as well, in order to poll on each iteration
until a set point is received or the cycletime is reached.

Two debug messages are added describing the previous functionality, and
the level of 2 previous log messages are changed, in order to print more
adequate messages on the console.