Add loss of UGV detection on 4 stepsand improve plot
The detection loss is aimed to reproduce the real phenomena on the
UviSpace Localization nodes. It is achieved setting the measurement
noise to a very high value.

A legend is added to the figure. It is added on the first iteration
because on each iteration 4 new plots are created. Thus, if it is
created at the end, as usual, it would add 4 legend entries for each
program iteration.

Finally, the ImportError is evaluated on the kalmanfilter import, as
this instruction is executed before the importing of settings module.