Add a function in messenger for getting the battery SOC
the function will be called at regular time intervals, defined by the
new variable soc_read_intervals, set by default to 5 seconds.

At the moment, there is a catch: As the listener read instruction is
indefinitely waiting for an incoming message, the new function will only
be accessed after a speed set point is received.

In the future, the battery state-of-charge should be published through
a socket, in order to deal with it by the controller module.