Unpack the SOC data correctly, and change log levels
The incoming state-of-charge (SOC) is received on 4 bytes (C chars), but
the useful data is always stored on the last 2 bytes, as it will never
be bigger than 100. Thus, only the las 2 bytes are interpreted in order
to get and print the battery SOC.

The log levels in serialcomm have been degraded to DEBUG, as they offer
no relevant information when a normal execution is performed.