Change the variable types for keeping coherency with incoming data
The external PCB connected via I2C to the Arduino board sends data with
unsigned int format. Thus, for correctly reading this data, the used
variable *soc* must be of the same type, instead the previously used
char type.

The process_message function has to separate the read data onto unitary
chars (1-byte length) before sending them through serial port. This way,
the length of the output message will be always constant. If not, when
an element is bigger than 10 the size will be 2, and if it is lower than
10, the size will be 1, making it difficult to set the message length.