Improve code and fix errors
Deleted geometric calculations in pathtracker module and solution them
in robot module. pathtracker module only calculate speed for UGV.

Reverted change to 2 blank lines between module functions.

Changed arctan by arctan2 in beta angle calculation.

Improved name for variables of coefficients, and polynomial solver
objects for a forward movement and turn movement in many modules.

It saves the route traveled by the UGV and the ideal route of this. The
data is shown in a graph when shutdown navigator. (This function is an
adaptation of a before function).

Speedstudy module obtain coefficients to modelrobot.cfg. This
configuration file can be modified by the user.

Corrected main routine of multiplecamera.

Changed name for logger: 'speed' is called 'speedstudy'.

Fixed many minor errors.